back-end son islemler

+++ B2BBackEnd/Business/Repositories/ProductRepository/ProductManager.cs
@@ -6,6 +6,9 @@ using System.Threading.Tasks;
 using Business.Repositories.ProductRepository;
 using Entities.Concrete;
 using Business.Aspects.Secured;
+using Business.Repositories.BasketRepository;
+using Business.Repositories.OrderDetailRepository;
+using Business.Repositories.OrderRepository;
 using Business.Repositories.PriceListDetailRepository;
 using Business.Repositories.ProductImageRepository;
 using Core.Aspects.Validation;
@@ -13,6 +16,7 @@ using Core.Aspects.Caching;
 using Core.Aspects.Performance;
 using Business.Repositories.ProductRepository.Validation;
 using Business.Repositories.ProductRepository.Constants;
+using Core.Utilities.Business;
 using Core.Utilities.Result.Abstract;
 using Core.Utilities.Result.Concrete;
 using DataAccess.Repositories.ProductRepository;
@@ -25,12 +29,16 @@ namespace Business.Repositories.ProductRepository
         private readonly IProductDal _productDal;
         private readonly IProductImageService _productImageService;
         private readonly IPriceListDetailService _priceListDetailService;
+        private readonly IBasketService _basketService;
+        private readonly IOrderDetailService _orderDetailService;
 
-        public ProductManager(IProductDal productDal, IProductImageService productImageService, IPriceListDetailService priceListDetailService)
+        public ProductManager(IProductDal productDal, IProductImageService productImageService, IPriceListDetailService priceListDetailService, IBasketService basketService, IOrderService orderService, IOrderDetailService orderDetailService)
         {
             _productDal = productDal;
             _productImageService = productImageService;
             _priceListDetailService = priceListDetailService;
+            _basketService = basketService;
+            _orderDetailService = orderDetailService;
         }
 
        // [SecuredAspect("admin,product.add")]
@@ -57,10 +65,19 @@ namespace Business.Repositories.ProductRepository
         [RemoveCacheAspect("IProductService.Get")]
         public async Task<IResult> Delete(Product product)
         {
+            IResult result = BusinessRules.Run(
+                await CheckIfProductExistToBaskets(product.Id),
+                await CheckIfProductExistToOrderDetails(product.Id)
+                );
+            if (result != null)
+            {
+                return result;
+            }
+
             var images = await _productImageService.GetListByProductId(product.Id);
             foreach (var image in images)
             {
-             var result=  await _productImageService.Delete(image);
+             await _productImageService.Delete(image);
 
             }
             var priceListProduct=await _priceListDetailService.GetListByProductId(product.Id);
@@ -93,6 +110,29 @@ namespace Business.Repositories.ProductRepository
         public async Task<IDataResult<Product>> GetById(int id)
         {
             return new SuccessDataResult<Product>(await _productDal.Get(p => p.Id == id));
+        }
+
+        public async Task<IResult> CheckIfProductExistToBaskets(int productId)
+        {
+            var result = await _basketService.GetListByProductId(productId);
+            if (result.Count()>0)
+            {
+                return new ErrorResult("Silmeye çalýþtýðýnýz ürün sepette bulunuyor.");
+            }
+
+            return new SuccessResult();
+
+        } 
+        public async Task<IResult> CheckIfProductExistToOrderDetails(int productId)
+        {
+            var result = await _basketService.GetListByProductId(productId);
+            if (result.Count()>0)
+            {
+                return new ErrorResult("Silmeye çalýþtýðýnýz ürünün sipariþi bulunuyor.");
+            }
+
+            return new SuccessResult();
+
         }
 
     }
